moved the job-ad-rules.ts to the shared jobs module
